+++ python/paddle/fluid/evaluator.py
@@ -306,25 +306,21 @@ class DetectionMAP(Evaluator):
     1. calculate the true positive and false positive according to the input
         of detection and labels.
     2. calculate mAP value, support two versions: '11 point' and 'integral'.
-      11point: the 11-point interpolated average precision.
-      integral: the natural integral of the precision-recall curve.
 
     Please get more information from the following articles:
       https://sanchom.wordpress.com/tag/average-precision/
       https://arxiv.org/abs/1512.02325
 
     Args:
-        input (Variable): LoDTensor,The detection results, which is a LoDTensor with shape
+        input (Variable): The detection results, which is a LoDTensor with shape
             [M, 6]. The layout is [label, confidence, xmin, ymin, xmax, ymax].
-        gt_label (Variable): LoDTensor, The ground truth label index, which is a LoDTensor
-            with shape [N, 1].The data type is float32 or float64.
-        gt_box (Variable): LoDTensor, The ground truth bounding box (bbox), which is a
+        gt_label (Variable): The ground truth label index, which is a LoDTensor
+            with shape [N, 1].
+        gt_box (Variable): The ground truth bounding box (bbox), which is a
             LoDTensor with shape [N, 4]. The layout is [xmin, ymin, xmax, ymax].
-            The data type is float32 or float64.
-        gt_difficult (Variable|None): LoDTensor, Whether this ground truth is a difficult
+        gt_difficult (Variable|None): Whether this ground truth is a difficult
             bounding bbox, which can be a LoDTensor [N, 1] or not set. If None,
-            it means all the ground truth labels are not difficult bbox.The
-            data type is int.
+            it means all the ground truth labels are not difficult bbox.
         class_num (int): The class number.
         background_label (int): The index of background label, the background
             label will be ignored. If set to -1, then all categories will be
@@ -334,9 +330,11 @@ class DetectionMAP(Evaluator):
         evaluate_difficult (bool): Whether to consider difficult ground truth
             for evaluation, True by default. This argument does not work when
             gt_difficult is None.
-        ap_version (str): The average precision calculation ways, it must be
+        ap_version (string): The average precision calculation ways, it must be
             'integral' or '11point'. Please check
             https://sanchom.wordpress.com/tag/average-precision/ for details.
+            - 11point: the 11-point interpolated average precision.
+            - integral: the natural integral of the precision-recall curve.
 
     Examples:
         .. code-block:: python
